handle class names with teact
