+++ web/src/scss/styles.scss
@@ -1,0 +1,2 @@
+// Import all of Bootstrap's CSS
+@import "~bootstrap/scss/bootstrap";
